+++ src/controllers/CkeditorController.php
@@ -12,9 +12,11 @@ use craft\ckeditor\Field;
 use craft\elements\Asset;
 use craft\fieldlayoutelements\CustomField;
 use craft\web\Controller;
+use Throwable;
 use yii\web\BadRequestHttpException;
 use yii\web\NotFoundHttpException;
 use yii\web\Response;
+use yii\web\ServerErrorHttpException;
 
 /**
  * CKEditor controller
@@ -88,6 +90,40 @@ class CkeditorController extends Controller
             'cardHtml' => $cardHtml,
             'headHtml' => $view->getHeadHtml(),
             'bodyHtml' => $view->getBodyHtml(),
+        ]);
+    }
+
+    /**
+     * Duplicates a nested entry and returns the duplicate’s ID.
+     *
+     * @return Response
+     * @throws BadRequestHttpException
+     * @throws ServerErrorHttpException
+     * @since 4.0.0
+     */
+    public function actionDuplicateNestedEntry(): Response
+    {
+        $entryId = $this->request->getRequiredBodyParam('entryId');
+        $siteId = $this->request->getBodyParam('siteId');
+        $entry = Craft::$app->getEntries()->getEntryById($entryId, $siteId, [
+            'status' => null,
+            'revisions' => null,
+        ]);
+
+        if (!$entry) {
+            throw new BadRequestHttpException("Invalid entry ID: $entryId");
+        }
+
+        try {
+            $newEntry = Craft::$app->getElements()->duplicateElement($entry);
+        } catch (Throwable $e) {
+            return $this->asFailure(Craft::t('app', 'Couldn’t duplicate {type}.', [
+                'type' => $entry::lowerDisplayName(),
+            ]), ['additionalMessage' => $e->getMessage()]);
+        }
+
+        return $this->asJson([
+            'newEntryId' => $newEntry->id,
         ]);
     }
 }